Add simple indexed texture
This is mostly intended as an example. In reality textures should be
constructed from buffer data received from the server. This also ignores
chunking of any kind, and makes no attempt to consider templates either.

This doesn't use an optimized format for the board data, such as GL_R8
or GL_LUMINANCE since the former didn't immediately just work and the
latter is supposedly deprecated. Ideally, such a format should be used
since it would save significant amounts of VRAM (as much as 75%).