Use 1 channel texture for canvas
In theory this reduces VRAM usage by 75%. Unlike usual, I actually
checked this and it does actually do that, so that's pretty neat.

The `LUMINANCE` texture format is apparently deprecated in webgl 2 but
it's the only format that I've found that consistently works. Webgl 2
has `R8` but that's webgl 2 only and also didn't work last time I
tested.

Perhaps there's some room for improvement in future, but this code
shouldn't stick around for long anway and just serves as proof that this
is a desirable optimization.