Implement simple chunk loading and unloading
This allows for practically infinite canvas sizes. The default
configuration in this commit sets the shape to
`[[500, 500], [500, 500]]` which is 62,500,000,000 pixels in total size.
This is 62,500 times the size of the standard 1000 by 1000 canvas.

It does this while requiring no more than a few megabytes of vram in use
at any given time. Well, at least in theory. In practice it seems that
at least my configuration of Firefox doesn't release used VRAM until the
window is closed. Everything worked fine in chromium for me at least.