Implement heatmap
This is seeded with random data, but illustrates the rendering quite
well. The new model of providing the client with timestamps rather than
a precalculated heatmap is a lot more flexible and allows for things
like realtime duration control and (admittedly imperfect) time-based
playback.

The precise layering of the template and the heatmap is a bit tricker
here than in pxls original situation. I've implemented the cheapest and
easiest solution for layering the template "behind" the heatmap in that
it is dimmed the same as the main canvas but will still render on top if
it. Actually rendering the heatmap above the template would require an
additional renderpass for each canvas piece which I'd like to avoid.