Abstract backend for sector data
This drops heatmap support temporarily because I just want to commit
this while it's working. What *is* working which was rather troublesome
to do is proper shape indexing. To illustrate this, a canvas of shape
`[[4, 4], [x, y]` for sectors of size (x, y) has a layout of:

    0  1  2  3
    4  5  6  7
    8  9  10 11
    12 13 14 15

But a canvas of shape `[[2, 2], [2, 2], [x, y]]` has a layout of:

    0  1  4  5
    2  3  6  7
    8  9  12 13
    10 11 14 15

The new code should be able to handle all this seamlessly for any shape.