Restructure and implement persistent and url stores
This sets things up with a more structured approach based on a backing
model for each component as well as puts the ground work in for a login
system.

It also makes use of zod for parsing, which isn't quite as nice as
serde's derive, but goes a long way to making parsing easier and
simpler.