Rewrite much of rendering, implement board choice and placing
There's a lot in here, but it's pretty rough in some areas so there's
plenty of cleanup to do. On the rendering side, the render preparation
side is now a lot cleaner and straightforward while on the other end we
now take advantage of instanced rendering to render up to 4 tiles in one
render call. The same goes for templates which can now be rendered 7 at
a time. I have yet to check if this improves performance at all given
the overhead of complexity and dynamic texture bind access requiring
conditions but I'm banking on glsl being faster than JS.

On the network side, this now implements fetching of board info for
multiple canvases as well as a simple ui to select which one to use.
This is helpful in debugging, but will likely be set to use the
default board in most cases.

The reticule here is rendered in dom rather than webgl, meaning it can
be slightly off from the displayed board. This can probably be corrected
for a little eventually — input handling in general needs a big rework
pass — but for now it's good enough.