Implement heatmap controls
The duration ranges are fairly flexible at the moment, but run on a
fixed formula. It may prove valuable in future to come up with a more
powerful system to set these parameters.

In addition, I am considering adding a setting to hide the offset range
behind an "advanced heatmap control" toggle to keep the default
interface cleaner, given most poeple will likely not need the temporal
offsetting.