Refactor parsing and caching
This makes the business code a lot simpler and easier to write at the
cost of the targetted code getting a bit more complicated in the
background. There's definitely room to more iteratively improve that in
future, but this gets it out of the way for now.

This also minorly improves permissions and extension handling since some
preliminary work got tangled up in this commit, there's a lot more of
that still to do though.