Reload page on socket connection loss
This is a terrible solution akin to "turn it off and on again" but the
reality is it's simple to implement and works, unlike a required
refactor of all the state dependencies currently existing.

Perhaps if time permits, a more graceful failure and recovery mode can
be worked on but this solution has worked for pxls for a very long time
now, so it can work here in a pinch too.