Support touch events and refactor input
While it may seem innocuous, there's a lot of work here to provide a
very smooth input experience. I reworked it from the ground up to
support touch and consequently it should feel very solid. There's still
a fair bit of sketchyness with quick-placing and due to touch-caused
events behaving differently to pointer-caused events regarding event
source and target, but if possible that can be cleaned up later.

This is still missing a couple of things I'd like to add to camera
handling: notably some momentum on fast pan and also gradual recentering
when the soft-bounds are hit to produce bounce-like effects.