Improve experience for fast scroll-rate devices
Prior to this the camera could jump all over the place when scrolling
out to the minimum limit whenever usig the instant scroll. For trackpad
devices which send scroll events very frequently this resulted in a very
unpleasant experience. By throttling the physics to 50ms after the last
event, this largely solves the problem without much noticable delay to
other input methods, such as a mouse scrollwheel.