Improve behaviour when entire board fits inside screen
The grab resume freaks out a bit here. I think this is because it
doesn't quite have the information to compensate for both overlaps
affecting it at once. On a whim, I tried diving by two since the
overlaps values are sort of averaged when both are present and this
seems to have improved things (although not fixed them).

Given the complex nature of this code, and the recentering physics
usually at play making this difficult to notice, I'm going to accept
this solution for now lest I waste a week trying to perfect this single
minor thing.