Prevent periodic jitter when zoomed out
Despite the attempts to stop physics, the board not being totally
centered was resulting in the re-render from the one second timer
interval ticking the physics loop, which results in a slight force being
applied to the board in one direction or the other, disrupting physics
sleep.

By only applying forces when they are significant, similar to how
friction was already applied, the velocity remains at 0 and the board
comes to a nice rest and stays there.