Fix layout and input bugs when canvas is not fullscreen
While not exactly necessary, being able to resize and reposition the
canvas however desired with css is a very "nice to have" feature and
since it was mostly designed with that in mind, I decided to fully test
and implement it. This also has the benefit of eliminating the use of
layerX/layerY event properties which are nonstandard and caused buggy
behaviour if scrolling is to be used on the palette.