Condense palette and implement custom scrolling
Since the default scrolling conflicts with quick-placing on mobile, we
regrettably need to implement our own scrolling. This basically takes a
lot of the code from the canvas fling and uses it to create a natural
feeling scroll on the palette, but needing to do this in the first place
is far from ideal.