Prevent context menu on canvas
Since the unique context menu items here to save the image and open it
in a new tab are misleading and don't work anyway without enabling webgl
buffer preservation (in which case, it would still only show the
current screen rather than the whole canvas as one might like to
believe).

This also opens the door to the possibility of binding specific pointer
buttons in future, but for now all act similarly.