Fix zod chain parsing
The previous method of expecting `transform` to catch errors was
outright incorrect. This changes locations where `transform` was
used to use the `pipe` function instead. In addition, where transform
does need to be used this makes sure to correctly use the
RefinementContext parameter to chain internal parsers into the current
context via `safeParse`.