Improve template style rendering
The rendering at fault is the appearance of single pixel opaque lines at
the borders of pixels even when they should be transparent. I don't
remember the exact cause since I investigated it a while ago, but I do
recall that the culpret was filtering (min filtering it seems, as mag
was already nearest neighbor).

The downside to setting this is that we lose access to mipmaps which
would mitigate some artifacts at certain zoom levels. While less
distracting subjectively, neither option is ideal.