Implement Unwrap compontent to allow more flexible store usage
Svelte's $ syntax for stores can only unwrap them at the top level of a
component. Since we have quite a few nested store-promise chains, this
required carefully injecting each store as the root parameter of every
component that needs it and making a new subcomponent for every other
substore. With this new generic component, we can instead unwrap and
await these stores to provide each component with only the data is it
expecting, and we are no longer forced to create extra components just
to unwrap any substores.