Fix double socket connection
This was a really silly bug introduced when missed event playback was
introduced. The short of it is: rather than only creating a new socket
if we didn't already have one and removing the listener if we did, the
code just did both actions so long as the socket was already created.