Fetch token code from query instead of fragment
This is the default behaviour specified in oauth2. Of course, setting
the response_mode does allow returning this in the fragment but the
default is query and I can't find any recommentation as to which
response_mode values are required to be supported. Querying the
discovery endpoint should give a list, but supposing that fragment isn't
in that list, this will have to handle query anyway.

With that in mind, just do query parsing anyway. This also comes with
the benefit of additional scrutiny on the fragment updating code, which
will now no longer remove query params when updated.