Remove "basic" scope from authentication
This seems to have been a keycloak specific thing. The purpose seems to
have been the simplification of some values. Alarmingly, one of these
values is the "sub" claim which the spec states should be "never
reassigned".

Since the value is non-standard, the safest thing to do is remove it.
This does mean that anyone with an existing userbase on keycloak will
now have different sub values for every user, which would likely be a
significant issue. This is precisely why I imagine the spec indicates
that sub should never be reassigned, but given few people if any are
currently running this software, making the switch now is the most
painless option.